修改BUG

+++ service/src/main/java/com/bestomb/service/impl/MasterServiceImpl.java
@@ -6,6 +6,7 @@ import com.bestomb.common.exception.EqianyuanException;
 import com.bestomb.common.request.tombstone.master.MasterEditRequest;
 import com.bestomb.common.response.PageResponse;
 import com.bestomb.common.response.master.MasterBo;
+import com.bestomb.common.util.CalendarUtil;
 import com.bestomb.common.util.FileUtilHandle;
 import com.bestomb.common.util.SessionUtil;
 import com.bestomb.common.util.yamlMapper.SystemConf;
@@ -154,6 +155,9 @@ public class MasterServiceImpl implements IMasterService {
         master.setLifeIntroduce(masterEditRequest.getLifeIntroduce());
         master.setLastWish(masterEditRequest.getLastWish());
         master.setAge(masterEditRequest.getAge());
+        master.setCreater(masterEditRequest.getCreater());
+        master.setCreateTime(masterEditRequest.getCreateTime());
+        master.setDataTime(CalendarUtil.getSystemSeconds());
 
 //        if (!ObjectUtils.isEmpty(masterEditRequest.getPortraitFile())) {
 //            //纪念人头像上传
@@ -174,10 +178,12 @@ public class MasterServiceImpl implements IMasterService {
          * 目录结构：持久化上传目录/陵园编号/master_portrait/文件
          */
         String portraitPath = SystemConf.FILE_UPLOAD_FIXED_DIRECTORY.toString() + File.separator + masterEditRequest.getCemeteryId() + File.separator + "master_portrait";
-        master.setPortrait(portraitPath + File.separator + masterEditRequest.getPortraitName());
-        //将纪念人头像文件从临时上传目录移动到持久目录
-        String absoluteDirectory = SessionUtil.getSession().getServletContext().getRealPath("/");
-        FileUtilHandle.moveFile(absoluteDirectory + SystemConf.FILE_UPLOAD_TEMP_DIRECTORY.toString() + File.separator + masterEditRequest.getPortraitName(), portraitPath);
+        if(!StringUtils.isEmpty(masterEditRequest.getPortraitName())){
+            master.setPortrait(portraitPath + File.separator + masterEditRequest.getPortraitName());
+            //将纪念人头像文件从临时上传目录移动到持久目录
+            String absoluteDirectory = SessionUtil.getSession().getServletContext().getRealPath("/");
+            FileUtilHandle.moveFile(absoluteDirectory + SystemConf.FILE_UPLOAD_TEMP_DIRECTORY.toString() + File.separator + masterEditRequest.getPortraitName(), portraitPath);
+        }
 
         //持久化纪念人数据
         masterDao.insertSelective(master);
@@ -228,15 +234,19 @@ public class MasterServiceImpl implements IMasterService {
         master.setLifeIntroduce(masterEditRequest.getLifeIntroduce());
         master.setLastWish(masterEditRequest.getLastWish());
         master.setAge(masterEditRequest.getAge());
+        master.setCreater(masterEditRequest.getCreater());
+        master.setCreateTime(masterEditRequest.getCreateTime());
 
         //检查新老图片信息是否一致，不一致说明有更新
         if (!newProtrait.equals(oldPortrait)) {
-            //将纪念人头像文件从临时上传目录移动到持久目录
-            String absoluteDirectory = SessionUtil.getSession().getServletContext().getRealPath("/");
-            FileUtilHandle.moveFile(absoluteDirectory + SystemConf.FILE_UPLOAD_TEMP_DIRECTORY.toString() + File.separator + masterEditRequest.getPortraitName(), portraitPath);
-            //将纪念人旧头像文件从持久目录删除
-            FileUtilHandle.deleteFile(SessionUtil.getSession().getServletContext().getRealPath("/") + oldPortrait);
-            master.setPortrait(newProtrait);
+            if(!StringUtils.isEmpty(masterEditRequest.getPortraitName())){
+                //将纪念人头像文件从临时上传目录移动到持久目录
+                String absoluteDirectory = SessionUtil.getSession().getServletContext().getRealPath("/");
+                FileUtilHandle.moveFile(absoluteDirectory + SystemConf.FILE_UPLOAD_TEMP_DIRECTORY.toString() + File.separator + masterEditRequest.getPortraitName(), portraitPath);
+                //将纪念人旧头像文件从持久目录删除
+                FileUtilHandle.deleteFile(SessionUtil.getSession().getServletContext().getRealPath("/") + oldPortrait);
+                master.setPortrait(newProtrait);
+            }
         }
 
         //持久化纪念人数据
